More Re-Organizing
moved a few other files that were saved in wrong spots. deleted the "vacation folder" as the only file in there was moved to "gitcommit Homework" folder. it was an image for that assignment. Moved some notes that were filed in the the home work folder to the notes folder. A screen grab for solo learn that was saved in the homework folder now is in the SoloLearn folder.